feats: show enemy stats on quest page and move level/exp calc to separate util file
